requirements.txt
sherpa-onnx
sounddevice
numpy
pynput
rumps
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
